+++ src/mgr/MgrPyModule.h
@@ -45,13 +45,14 @@ private:
   const std::string module_name;
   PyObject *pClassInstance;
   PyThreadState *pMainThreadState;
+  PyThreadState *pMyThreadState = nullptr;
 
   std::vector<ModuleCommand> commands;
 
   int load_commands();
 
 public:
-  MgrPyModule(const std::string &module_name, PyThreadState *main_ts);
+  MgrPyModule(const std::string &module_name, const std::string &sys_path, PyThreadState *main_ts);
   ~MgrPyModule();
 
   int load();
